requirements.txt
requests>=2.31.0
python-docx>=1.1.0
python-dateutil>=2.9.0
python-dotenv>=1.0.1
